waklin list and attendess modified
